Refactoring the Code
Cleaning methods up a bit, so we can starting implementing our HTML Layout
for better user control of the layout for the PDF book. Also, slimmed down
the arguments for passing to methods. Instead, the book directory, will be
the layout for how the HTML and PDF books are created.