Added Syntax Highlighting
Added highlighting of code blocks that will support multiple
languages within one book. The trick is adding ::language:: at
the top of the code block. This will tell the parser what language
the following code is and highlight it appropriately.

The ::language:: will be stripped out of the HTML & PDF Versions
of the book.