Refactored to Classes
Makes more sense to have an HTML::Book class and a PDF::Book class.
This provides two different versions, while the PDF version will
ride on the back of the HTML version.

Added in a readme document to explain what MarkaPDF is, some of the
ways to use it and the future plans for this library.